+++ src/api/java/com/codetaylor/mc/artisanworktables/api/internal/reference/Tags.java
@@ -10,6 +10,7 @@ public final class Tags {
   public static final String TAG_UPGRADE_ENERGY_USAGE = "EnergyUsage";
   public static final String TAG_UPGRADE_FLUID_CAPACITY = "FluidCapacity";
   public static final String TAG_UPGRADE_ENERGY_CAPACITY = "EnergyCapacity";
+  public static final String TAG_UPGRADE_AUTO_EXPORT_ITEMS = "AutoExportItems";
 
   private Tags() {
     //