+++ src/main/java/com/codetaylor/mc/artisanworktables/modules/automator/ModuleAutomator.java
@@ -76,10 +76,14 @@ public class ModuleAutomator
     @GameRegistry.ObjectHolder(ItemUpgrade.NAME_ENERGY_CAPACITY)
     public static final ItemUpgrade UPGRADE_ENERGY_CAPACITY;
 
+    @GameRegistry.ObjectHolder(ItemUpgrade.NAME_AUTO_EXPORT_ITEMS)
+    public static final ItemUpgrade UPGRADE_AUTO_EXPORT_ITEMS;
+
     static {
       UPGRADE_SPEED = null;
       UPGRADE_FLUID_CAPACITY = null;
       UPGRADE_ENERGY_CAPACITY = null;
+      UPGRADE_AUTO_EXPORT_ITEMS = null;
     }
   }
 
@@ -166,6 +170,7 @@ public class ModuleAutomator
     registry.registerItem(new ItemUpgrade(), ItemUpgrade.NAME_SPEED);
     registry.registerItem(new ItemUpgrade(), ItemUpgrade.NAME_FLUID_CAPACITY);
     registry.registerItem(new ItemUpgrade(), ItemUpgrade.NAME_ENERGY_CAPACITY);
+    registry.registerItem(new ItemUpgrade(), ItemUpgrade.NAME_AUTO_EXPORT_ITEMS);
 
     //noinspection unchecked
     this.registerTileEntities(
@@ -188,7 +193,8 @@ public class ModuleAutomator
       ModelRegistrationHelper.registerItemModels(
           Items.UPGRADE_SPEED,
           Items.UPGRADE_FLUID_CAPACITY,
-          Items.UPGRADE_ENERGY_CAPACITY
+          Items.UPGRADE_ENERGY_CAPACITY,
+          Items.UPGRADE_AUTO_EXPORT_ITEMS
       );
     });
   }